fix: RUN-874: Fix the return type of some System APIs (#785)
Currently some System APIs smaller range of result than specified:
- ic0_canister_cycle_balance (returns i64 instead of u64)
- ic0_msg_cycles_available (returns i64 instead of u64)
- ic0_msg_cycles_refunded (returns i64 instead of u64)
- ic0_mint_cycles (returns i64 instead of u64)
- stable64_size (returns i64 instead of u64)
- stable_size (returns signed i32 instead of unsigned u32)

If the result doesn't fit into the narrow range, then the System APIs
trap.
This PR fixes it and changes the System API return the correct range.

Regression tests are added for all System APIs except for
`stable[64]_size()`
that can never return values that don't fit in `i32` (because the result
is in Wasm pages).