feat(ckerc20): Use EVM-RPC canister to call `eth_getTransactionCount` (#792)
([XC-164](https://dfinity.atlassian.net/browse/XC-164)): Continue the
migration initiated in
[MR-19972](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/19972)
towards using the EVM-RPC canister to interact with the Ethereum
JSON-RPC providers. This PR focuses specifically on the
`eth_getTransactionCount` endpoint.

This endpoint needs to be handled slightly differently than the other
ones because the same type `TransactionCount` is handled differently
depending on the block height for which it was queried:
1. `Finalized`. In that case, equality between multiple
`TransactionCount` for the same Ethereum address is expected.
2. `Latest`. In that case, equality between multiple `TransactionCount`
for the same Ethereum address cannot be expected and so the minimum is
taken.