fix: Subnet splitting with canister snapshots
This PR fixes subnet splitting in the presence of canister snapshots.

During subnet splitting, we first load a state, prune all data only
relevant to the other split, and then trigger the checkpointing code in
order to write a new checkpoint to disk. With canister snapshots, we now
need to make sure that we call `flush_page_maps`. This step is part of
the full checkpointing procedure, but was omitted during subnet
splitting as it was a noop. However, this function is now also handling
canister snapshots, particularly it is responsible for removing
deleted snapshots. Hence we need to call it during subnet splitting.

This PR also extend the subnet splitting tests to include snapshots. As
canister snapshots are not possible for canisters without installed code,
the test is updated to install code for all canisters in the subnet.
This affects the expected set of files that this test checks for.