refactor: Use ic_cdk::api::time for ingress message validator crate (#802)
This PR switches the ingress message validator crate from using
`dfn_core::now()` (`dfn_core` is deprecated), to use the public Rust CDK
instead.

`dfn_core` was written in way that would return some `Time` when
compiled to native code which would allow the doc test to run
successfully. In contrast, the CDK assumes that it compiles only to
`Wasm` targets so we need to add some `cfg` directives to ensure the
code can compile in either native or `Wasm` targets.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>