refactor: Switch some test canisters in rust_canisters to use public CDK (#805)
This PR updates some of the test canisters included in `rust_canisters`
to use the public CDK instead of the deprecated internal one (`dfn_core`
and friends). The change is made for a set of canisters that were easy
to port over. There are some more cases that are more nuanced and will
happen in individual follow-ups.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>