chore: Update visibility of `dfn_*` libraries (#804)
This PR updates the visibility of `dfn_*` libraries to only the packages
that really need them.

`dfn_*` libraries in `rust_canisters` are considered the "internal" Rust
CDK which is long deprecated. We eventually plan to remove it but we
cannot yet as different parts of the code still depend on them.

The first step is to disallow usage of these libraries outside of their
current users to prevent them from spreading out even more. The
restrictions are as strict as they can reasonably be without having to
list a gazillion of packages (e.g. `//rs/nns` is whitelisted for all its
packages as the majority needs them anyway).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>