feat(ic-admin): Support sending update_canister_settings proposals through ic-admin (#789)
# Why

The `update_canister_settings` proposal type is added to NNS Governance
(currently, for testing), and we should make it possible for someone to
propose using ic-admin.

# What

Define a subcommand to propose to update canister settings. For updating
controllers, clap doesn't seem to support specifying a `Some(vec![])`,
therefore another argument `--remove-all-controllers` is added for the
case where `controllers` needs to be updated to be an empty list.

# Tests

Tested end-to-end locally with different arguments