test(consensus): Only consider summary blocks in CUP exhaustive test   (#797)
CUPs cannot hold blocks with data payloads, so it's not necessary to
generate exhaustive test cases with such payloads.

This has an advantage that
[DataPayload](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/types/types/src/consensus/payload.rs?L18:12-18:23)
won't have to be backwards compatible anymore, and making changes to the
structure won't break the `cup_compatibility_test`.

Note that since https://github.com/dfinity/ic/pull/795 we still test the
serialization->deserialization roundtrip of the `DataPayload`.