test(cloner-canister): Add manual system test to stress test subnet by installing order of 100k canisters (#798)
As part of stress testing the IC with lower notarisation delays, one
scenario we wanted to test is subnets that have large number of
canisters installed. This PR adds a system test that installs a large
number of canisters, to allow for reproduction of this stress test
scenario.

The PR:
- Adds a manual system test that spins up a configured amount of
canisters. 125k by default.
- Added a `cloner_canister` to our suite of canisters available for
system tests. This canister simply has an update call endpoint to spawn
a configured number, 250 by default, canisters.

---------

Co-authored-by: Bas van Dijk <bas@dfinity.org>