fix: Store the version of the snapshot that was loaded in canister history (#837)
This fixes a discrepancy between the interface spec and the replica wrt
the canister version stored in the `load_snaphot` history entry.

The change details of the history entry used to store the version of the
canister after loading the snapshot (this already appears in the history
entry anyway). However, it should contain the version of the canister at
the time that the snapshot was taken -- this along with the time taken
and the snapshot id, fully describes the snapshot that was loaded back
into the canister.