fix: disable node exporter netlink metrics collection (#826)
Recently the node exporter was upgraded from 1.2 to 1.8. With this
change the `arp` and `netdev` collector switched to using `netlink` to
collect metrics. The current selinux policy on the guestos does not
allow the node exporter process to open netlink socket and therfore the
the `arp` and `netdev` metrics are currently missing.

This MR disables the netlink metrics scarping by switching back to
scaping `/proc/net/dev` as was done by default before version 1.4.