feat(IDX): run execution tests on namespace
This moves some executions tests (known to be flaky on old Intel macs)
to the namespace runners. The namespace workflow is also changed to not
cancel current jobs on `master` so that we can check all commits built
successfully on `master`. Finally the `test_macos_master` flag is
renamed to `test_macos_slow`, which is more representative of what the
tests are, and less tied to the environment (branch names, etc).