chore(IDX): only upload build events on CI (#839)
Before this, build events were uploaded on every `build` & `test`
locally. When developing, `build` commands are usually incremental &
relatively fast, making the build event upload often take longer than
the build itself.

This moves the build event related config to a new config which can be
enabled with `--config=bes`, and makes the `--config=ci` inherit from it
so that it's still enabled on CI.