feat(ckerc20): Use EVM-RPC canister to call `eth_sendRawTransaction` (#836)
([XC-163](https://dfinity.atlassian.net/browse/XC-163)): Continue the
migration initiated in
[MR-19972](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/19972)
towards using the EVM-RPC canister to interact with the Ethereum
JSON-RPC providers. This PR focuses specifically on the
`eth_sendRawTransaction` endpoint.

One notable change of behaviour is that when the EVM-RPC canister is
used for `eth_sendRawTransaction`, all 3 providers (Ankr, Public Node
and Llama Modes) will be queried in parallel, whereas the minter uses a
minor optimization in querying all 3 providers sequentially until the
first ok result, since the sent transaction is signed and cannot be
altered.

[XC-163]:
https://dfinity.atlassian.net/browse/XC-163?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ