feat(nns): Redact private neuron info fields. (#842)
This feature is currently disabled in release builds, and enabled in
feature = "test" builds.

This also includes a tiny new library: ic_nervous_system_temporary. This
just makes it easier to test flag-protected features. It makes it pretty
easy to test the behavior both when the feature is enabled and when it
is disabled without the need to test two different builds. This new
library is in its own commit, which should make reviewing this PR
easier.