feat(registry): Add consistently-named alias `Registry.deploy_hostos_to_some_nodes` (#865)
This PR adds a new Registry function
`Registry.deploy_hostos_to_some_nodes` that works exactly like the
legacy `update_nodes_hostos_version` function but has a name that is
consistent with the new IC OS proposal naming convention (for details,
see [NNS Gov
proposal](https://dashboard.internetcomputer.org/proposal/129630) and
[the forum
discussion](https://forum.dfinity.org/t/bringing-clarity-to-icp-upgrade-proposals/29626)).

If and when the Registry is released with this API extension, it will
enable switching NNS Governance to call the new function with a
consistent name instead of the legacy one; after that, it will become
possible to obsolete the legacy function `update_nodes_hostos_version`
for good.