fix(IDX): don't cargo build rocksdb on Linux (#861)
This removes the `rocksdb_backend` cargo feature and instead use
`cfg(target_os="macos")` directly. This allows us to use
`[target.cfg(maocos).dependencies]` in the artifact pool's `Cargo.toml`
to ensure the `rocksdb` crate is only built on macOS (it's not used on
Linux but until now was built nonetheless when using cargo).