feat: EXC-1683: Include snapshot memory usage in canister's memory usage (#857)
A new field is added in `SystemState` that captures the memory usage of
all snapshots that belong to a canister. The field is updated whenever a
snapshot is taken or deleted and it's persisted in the checkpoints as
well.

The change is both backwards and forwards compatible as the new field
will be initialized to 0 and will remain 0 while the feature is still
disabled. Without any snapshots there's no functional difference as a
memory usage of 0 would be additionally included in the canister's
memory usage.

---------

Co-authored-by: Andriy Berestovskyy <berestovskyy@gmail.com>
Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>
Co-authored-by: Andriy Berestovskyy <andriy.berestovskyy@dfinity.org>