fix: CON-1321 Make the orchestrator recover from having a broken local CUP (#572)
The orchestrator periodically persists the latest CUP of the subnet and
triggers replica version upgrades. In case of an exceptional version
incompatibility, it may happen that the local CUP can no longer be
de-serialized on the new version.

Currently, if the orchestrator fails to deserialize the local CUP, then
— in contrast to the incompatibility of the execution state — we would
not be able to do a normal subnet recovery because the orchestrator
itself is non-functional. Instead we would have to do a recovery on
failover nodes, and then re-provision the broken original nodes.

This PR allows the orchestrator to recover from broken local CUPs using
a normal subnet recovery targeting the original nodes. The observation
is that, while the full CUP in its deserialized form is required to
start the _replica_, the _orchestrator_ itself only requires insight
into a few fields of the CUP to remain functional. By only deserializing
these specific fields, the orchestrator may therefore continue to
function, even if other parts of the CUP cannot be deserialized.

Consequently, the solution proposed here only improves the situation as
long as fields required by the orchestrator (namely, NiDkgId and block
height) remain deserializable even for otherwise broken local CUPs.

This requires two major changes:
1. In order to find the "latest" CUP of the subnet, the orchestrator
needs to determine which subnet it is currently on by inspecting the
local CUP. If the local CUP cannot be deserialized, we instead try to
extract the subnet ID from the rust "proto" struct directly, by only
deserializing the NiDkgId.

2. Afterwards, when determining the new "latest" CUP, we do so by
comparing the height of the CUPs from their rust "proto" structs
directly, as opposed to deserializing them in their entirety. This means
that we will prefer a newer local CUP over the latest registry CUP, even
if the local CUP cannot be deserialized. This is to avoid falling back
to outdated registry CUPs. Then, if the selected "latest" CUP indeed
cannot be deserialized, we will return an error and retry until we find
a higher registry CUP.

Additionally, we adapt one of the hourly subnet recovery tests (no
upgrade, with tECDSA), to exercise the new path. This is done by halting
the subnet and then corrupting the local CUPs of all nodes on the
subnet. Afterwards we attempt to recover the subnet as usual.

**Note** that in this failure scenario, for now we still wouldn't be
able to deploy read-only access to the subnet while the local CUP is
broken. This is because in order to update SSH access, the
orchestrator's SSH component needs to know the subnet ID, which it will
currently only receive if the upgrade component returns it without
error. However, this is fine as in this recovery scenario, we shouldn't
need to download and replay any blocks on top of the latest state
anyway, as there is no way that the replica was even started with a
broken CUP.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>