chore(IDX): consolidate lock-generate and bazel-pin (#886)
This consolidates the `lock-generate` and `bazel-pin` scripts, making
the former use the latter instead of duplicating the logic.

The `bazel-pin` script is also updated to have more flexible argument
handling, and is updated to support a `--force` argument. Unless
`--force` is provided, the `bazel-pin` script will try to infer if a
repin is indeed necessary, and if not, will not try to repin.

The `lock-generate` script is update to remove references to GitLab.

---------

Co-authored-by: Marko Kosmerl <marko.kosmerl@dfinity.org>