refactor: Remove PageMapType::get_mut
We need to iterate over all PageMaps at various places in the
StateManager. We do this typically with the help of the PageMapType
type, which encapsulates the details of where in the state we have
PageMaps. However, this pattern is awkward when we need to modify the
PageMap, and need to take into account canister snapshots.

This previously led to a get_mut function with some unimplemented!()
macros. These macros can turn bugs into runtime crashes.

Instead, this PR now removes the get_mut function and anywhere
where we need to iterate over all PageMaps and mutate them we now
instead do this without relying on the PageMapType encapsulation.