refactor: New test setup for stream handler tests (#564)
This provides a new framework for `StreamHandler` tests and refactors
the tests to use it.

`StreamHandler` tests all work with the same pattern:
- A setup stage where loopback/outgoing streams are populated with
messages and signals. Same for incoming stream slices.
- A stage where the expected state changes are produced, i.e. we expect
some messages to be inducted, others are rerouted or gc'ed and so forth.
- A call to a function of the `StreamHandler` using the state produced
in stage 1.
- A comparison of the resulting state after this call to the expected
state produced in stage 2.
- A check on the metrics. What was inducted? Which critical errors were
raised and which were not raised and so on.

The previous setup was inconsistent with the `CallbackId's` since they
were not registered properly. The new setup takes a number of 'configs'
to produce messages in streams and stream slices together with a
canister state, such that all `CallbackId's` are properly registered and
all the messages can be inducted at will.

Once a state prepared for the test has been produced in this way, an
expected state can be produced by cloning the initial state, generating
new streams by directly cloning messages from streams in the initial
state or generating reject responses (in stage 2).

Some other refactoring was added to remove boiler plate and make the
tests more readable and to the point. On average tests are 20% to 30%
shorter than before.