feat: [MR-581] Add functionalities of creating and removing unverified checkpoint markers (#657)
This PR adds a marker file whenever a tip or a state sync scratchpad is
promoted to an official checkpoint. This marker file is only removed
once the checkpoint has been successfully loaded.

If the loading fails, the replica will crash, leaving the marker file
within the checkpoint directory.  Upon restarting, any checkpoints
containing a marker file will be archived and ignored, preventing the
system from entering a continuous crash loop due to repeatedly
attempting to load a corrupt checkpoint.

With the marker file, we don't need to load checkpoint the synced
checkpoint twice at the end of state sync.

Writing the marker file is limited to `CheckpointLayout` with
`WritePolicy`.

---------

Co-authored-by: Stefan Schneider <31004026+schneiderstefan@users.noreply.github.com>