fix(IDX): remove darwin container check (#950)
This removes the custom `darwin_puller` that we set on `container_pull`,
which was used to make the container pulls fail on Darwin.

The non-existent label caused issues with `bazel query` so they are
removed. If the (now removed) check seems necessary in the future we'll
figure out a better way to implement it.