test(network-simulation): Increase transmission control buffers (#908)
This MR refactors the `simulate_network` module to support more
functionality, and refactors all system tests in the `consensus`
directory to use this module.

The previous implementation of setting simulated latencies in the
consensus system tests were faulty due to the transmission control
buffers overflowing, leading to packets being dropped. This is now
fixed.

As a consequence, the
[SUCCESS_THRESHOLD](https://github.com/dfinity/ic/blob/d253a7a508d37d8922ea7968e98f52e8c139edf8/rs/tests/consensus/consensus_performance.rs#L89)
in `consensus_performance.rs` can be removed as all requests now
succeed.