refactor(nns): Cleanup NNS Governance API type definitions (#961)
We previously auto-generated these, but the auto-generated added a lot
of cruft because the files were not meant to be hand-edited. Now that
we're hand-editing them, we can simplify some of that.

This PR does **not** actually change any of the types in a practical
sense, it just reorganizes how they refer to the types of their fields