chore: CON-1356 Make `Cannot report master public key changed metric` warning less noisy (#986)
This warning was previously logged whenever a node joins an
ECDSA/Schnorr subnet, or a node that isn't part of an ECDSA/Schnorr
subnet restarts, as these nodes never persisted a metric file
beforehand. We can make it less noisy and more meaningful by only
logging a warning if the file exists and then cannot be opened or read.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>