feat: ValidateEq derivation and annotation (#808)
When creating a checkpoint, we serialize the `ReplicatedState` on disk
and then load it back.
For further execution we use combination of both. If deserialized
`ReplicatedState` diverges from the original, it can lead to hard to
recover bugs, such as node diverging after state sync or machine
restart.
To find these divergences early, we can compare both states and trigger
a critical error in case of divergence. `ReplicatedState` does implement
`PartialEq` but that compares most of the data including up to hundreds
of GiB of pagemaps and queues. To have a fast comparison, we introduce
`ValidateEq` trait. It compares all the fields except explicitly
excluded ones, and provides a path to divergence for logging.

`ValidateEq` has an auto derive, that compares fields using PartialEq
one by one and reports which field has diverged. It's possible to
specialize treatment of individual fields to `Skip` (don't compare) or
`Recursive`. `Recursive` calls `ValidateEq` of the field. The
implementation doesn't allow to treat a field whose type implements
`ValidateEq` with default `PartialEq`, although the static assertion is
not bullet proof: it cannot catch ValidateEq in subfield or generics due
to Rust metaprogramming limitations.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>