feat(RUN-991): Set the default Wasm memory limit
This PR sets uninitialized `wasm_memory_limit` field in canister
settings to the default value, which is computed as follows:

- let `usage` be the Wasm memory size of the canister.
- let `halfway_to_max` be `(usage + 4GiB) / 2`.
- use `max(3GiB, halfway_to_max)` as the default value.

Note that if the field has already been initialized, then that value
is not overwritten.

For newly created canisters, the field is set to `3GiB`.

The initialization for existing canisters will happen at the end
of the first execution round after the replica upgrade.