feat(PSEC-2016): slack failover data store (#697)
This PR introduces the `SlackFindingsFailoverDataStore` which can be
used to manage specific findings in slack which can't be managed in
jira. The findings that should be managed in slack are determined by
`(repo, scanner, vuln_dependency_id)` and need to be configured in
[FAILOVER_FINDING_IDS](https://github.com/dfinity/ic/pull/697/files#diff-2b04e6b198b8af99581e09aa5b5519d0db3b7c94ef5e41eddfaa5a7b8afc7844R30).

In slack findings are managed by vulnerability in order to allow scaling
to large number of vulnerabilities. A vulnerability info message in
slack looks like this:

<img width="616" alt="Screenshot 2024-07-31 at 15 44 11"
src="https://github.com/user-attachments/assets/4e3aefba-9aa7-48c4-acac-882b76856498">


After every scan a summary message is posted that shows the number of
added/fixed/changed vulnerabilities and dependencies

<img width="455" alt="Screenshot 2024-07-31 at 15 44 18"
src="https://github.com/user-attachments/assets/4b312360-35af-4ac2-9f13-f2f61d30cd3a">

The integration supports pushing messages to different channels based on
the projects that are using a dependency. For each project a owning team
can be configured and for each team a separate slack channel can be
configured.

When the information of a vulnerability changes (e.g., description or
score), the corresponding slack message(s) are updated and a reply is
posted which lists the changed fields and their previous values. When a
vulnerability is fixed a `:fixed:` reaction is added to the slack
message.

Changes in dependencies, (e.g., added or removed dependencies) are
summarised in the scan results message mentioned above.