feat(sns): Reject new participants if the maximum number of required SNS neurons has been reached (#924)
This PR implements a more liberal enforcement for the max number of SNS
neurons. Currently, we enforce that the worst case is passable (by
validating the SNS init config at proposal submission time), while this
rejects some viable swap scenarios, including those that occurred in
production in the past. With this change, we can enforce the limits much
later, so the overly strict SNS init validation could be lifted (this
will be a follow-up PR).

For more details, see [this forum
thread](https://forum.dfinity.org/t/swaps-to-allow-at-most-100k-sns-neurons-for-direct-participants).