feat(nns-init): Add support for gzipped NNS canister wasms (#1024)
This commit modifies the `set_up_env_vars_for_all_canisters` function in
`lib.rs` to search for NNS canister wasms in the specified directory. It
checks for both `.wasm` and `.wasm.gz` files and sets up environment
variables accordingly. If a file is not found, it fails with an
assertion error, as usual.