feat(nns): add additional fields to get_monthly_node_provider_rewards (#1057)
This removes a no longer relevant field from the response
(use_registry_derived_rewards) that is only useful as part of the
proposal to reward node providers (which is also no longer used).

It also adds all the new fields to the response that are already
available and are helpful to clients in using this data.

The two types in the PR are compatible from a candid API perspective, so
migrating from one to the other is equivalent to removing an unneeded
field and adding the new fields.