test: Make payload size unique for all messages in StreamHandler tests. (#1055)
When tracking the `available_guaranteed_response_memory` GC-ed responses
are ignored. Therefore when comparing this to the computed value and
adjusting for this fact, only the right response will correctly
compensate. Without this change all responses have the same size and so
a test could pass but actually use the wrong message for memory checks.