chore(http-metrics): Increase number of buckets for certification time of messages (#1026)
In https://github.com/dfinity/ic/pull/1022 we will start tracking the
time it takes for a message to complete execution and its height to
become certified.

The previous buckets for the metric that tracks these times are coarse,
leading too high uncertainty in the histogram quantile calculation. This
metric is one of our best proxies to how fast the IC is preceived by end
users. We should therefore increase the number of buckets to get a
higher resolution view ingress call latencies.