feat(nns): Call new IC OS upgrade-related Registry functions instead of deprecated ones (#885)
This PR is the penultimate step towards consistency w.r.t. [this
table](https://forum.dfinity.org/t/bringing-clarity-to-icp-upgrade-proposls/29626#summary-of-the-planned-changes-6).
In particular, it changes the Registry functions that are called for the
following NNS proposals:
* `ReviseElectedGuestosVersions`
* `revise_elected_replica_versions` → `revise_elected_guestos_versions`
* `ReviseElectedHostosVersions`
  * `update_elected_hostos_versions` → `revise_elected_hostos_versions`
* `DeployHostosToSomeNodes`
  * `update_nodes_hostos_version` → `deploy_hostos_to_some_nodes`
* `DeployGuestosToSomeApiBoundaryNodes`
* `update_api_boundary_nodes_version` →
`deploy_guestos_to_some_api_boundary_nodes`

Additionally, this PR adjusts ic-admin and some system tests to form the
respective new payloads.