feat(sns): Expose the wasm_memory_limit in sns_canisters_summary's settings (#1054)
This is useful because it will allow us to tell at a glance which SNSes
already have their wasm_memory_limit set to 4gb and which are using the
default value (which will soon be changed to 3.5gb).

Also, I don’t understand
[this](https://github.com/dfinity/ic/blob/6ec7a4977268758d9a4eb7bf9166d24046baa111/rs/nervous_system/clients/src/canister_status.rs#L112-L113):
```
#[derive(CandidType, Debug, Deserialize, Eq, PartialEq, Clone, Default)]
pub struct DefiniteCanisterSettingsFromManagementCanister {
    pub controllers: Vec<PrincipalId>,
    pub compute_allocation: candid::Nat,
    pub memory_allocation: candid::Nat,
    pub freezing_threshold: candid::Nat,
    pub reserved_cycles_limit: candid::Nat,
    pub wasm_memory_limit: candid::Nat, // ← Why is this field not optional?
    pub log_visibility: LogVisibility,
}
```
I thought that `wasm_memory_limit` was optional, and if the field was
unset the runtime would use the default of 3.5GiB for that canister. But
it looks like the field the management canister returns is not optional.
I wonder if it's optional internally but the management canister just
returns 3.5GiB if it's not set? If so, does this mean there's no way to
tell the runtime to use the default value? cc @jasonz-dfinity since I
think you added this field. Either way, it should be safe to make it
optional on our end since `T` can be decoded to `Option T` in candid and
I don't think we encode this type, but I haven't changed it.

[← Previous PR](https://github.com/dfinity/ic/pull/1056)