fix(tests): Reduce Schnorr Message Size in upgrade_downgrade Tests (#1099)
In this test we send ingress messages to make sure that we can store and
read messages after an upgrade. At the same time we send a workload of
ECDSA/Schnorr messages to test the compatibility.

Because the Schnorr messages have a size of ~2MB, they sometimes block
other ingress messages from being included, meaning they will expire and
cause flakiness of the test.

As this isn't a performance test, reducing Schnorr messages to have a
more reasonable size shouldn't be an issue.