chore(ic-admin): Remove deprecated commands for subnet node add and remove (#1103)
change-subnet-membership should be used instead of separate remove-node
and add-node commands.
The `change-subnet-membership` command provides atomic node and removal,
so both operations can be performed in a single proposal.

Also, prominently show suggestions to use the DRE tool for submitting
this type of proposal, since the DRE tool calculates and optimizes
subnet decentralization instead of blindly adding or removing nodes.