feat(nns): Store `wasm_metadata` in SNS-W's stable memory (attempt #2) (#977)
This PR is the first step towards enabling the `SnsW.get_wasm_metadata`
endpoint in production. For that, we need a way to store the metadata in
the stable memory, as recomputing it each time is expensive (because
canister WASMs stored in SNS-W are often compressed).

Unlike the [first attempt](https://github.com/dfinity/ic/pull/956), this
PR makes it harder to represent disallowed states in the type system,
e.g., there cannot metadata that does not belong to any WASM stored in
SNS-W.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Carly Gundy <47304080+cgundy@users.noreply.github.com>
Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>
Co-authored-by: tim gretler <tim.gretler@dfinity.org>