feat(nns-release-tools): More ergonomic NNS & SNS upgrade proposals. (#1046)
Various changes that IMHO make NNS canister upgrade and SNS publish
proposals easier to generate, submit, and read.

- Instead of the title containing two giant opaque IDs, it says
something nice, like "Upgrade the Governance Canister to Commit abcd123"
- DRY_RUN environment variable.
- Instructions for how to verify WASM:
  - Mention pre-requisites. The instructions should now be complete.
  - Explain the steps.
- When asking whether the proposer is sure they want to submit the
proposal, print each ic-admin flag on its own line. (This is like how
you wouldn't put many statements in a Rust or C program on the same
line. Yes, it would be a more efficient use of screen space, but it is a
less efficient use of reader brain cycles, which is a far more precious
resource.)
- Use colors consistently (and hopefully, intuitively?):
  - prompts are green
- yellow is to specially call out important, but not critical
information
  - blue is for (ASCII art) line across the screen
- This might be controversial, but do not mention "uninteresting"
information in the proposal description. In particular, it no longer
mentions
- The WASM hash. This is only useful for verification, and for that
purpose, people should not be pulling this value from the description
anyway.
- Target canister ID. Instead, the description (like the title) only
mentions the canister's name. The reasoning here is the same: the ID can
be found elsewhere, is only useful for verification, and the description
is not a suitable source of such information.
- Because these two pieces of information are removed, verifying (at
proposal submission time) that they are consistent with other
information in the proposal (namely git commit ID, and canister name) is
no longer performed.
- Previously, if the internal `ic-admin` call failed, the proposal
submission script would still finish successfully. Now, our wrapper
scripts also fails in that case.
- More explicit (and celebratory) feedback to the user at the
conclusion.
- Link to additional information.
- Help strings are not so wide (max 80 char) anymore.
- Instead of downloading entire WASM during the proposal text
preparation, just make sure it exists (via `curl --head ...`).
(Previously, downloading the whole thing was necessary, since we wanted
to insert the WASM hash into the proposal text, but as mentioned
earlier, we got rid of this.)