feat(ic-admin): Print hashes rather than entire blobs when submitting InstallCode proposals (#1093)
# Why

Currently, the ic-admin prints the entire WASM when submitting the
`InstallCode` proposal. We want to only print hashes.

# What

When printing `InstallCode` proposals, we convert it into the "response"
API type which contains only hashes.

# Alternative

Ideally, we'd like to implement custom `Debug` trait. However, currently
it's impossible to do so while the proposal API type is still
`prost:message` which already derives `Debug` trait.