refactor(nervous_system): create ic-nervous-system-agent crate for interacting with the NNS and SNSes (#1087)
I thought it would make sense to have a crate where we would put various
helper functions for interaction with SNSes and the NNS. Right now, the
only rust helper functions I'm aware of are in some in
`//rs/nervous_system/tools/sync-with-released-nevous-system-wasms` for
interacting with SNS-W, so I moved just those for now. There are also
some that were developed for use in system tests, but they are not
trivial to extract so I haven't yet.

[Next PR →](https://github.com/dfinity/ic/pull/1091)