chore: update ingress watcher cancellation log level (#1133)
Currently, an
[error](https://github.com/dfinity/ic/blob/26d5f9d0bdca0a817c236134dc9c7317b32c69a5/rs/http_endpoints/public/src/call/ingress_watcher.rs#L250)
is logged if the ingress watcher is cancelled, but in production the
cancellation token is never used (since a
[fresh](https://github.com/dfinity/ic/blob/26d5f9d0bdca0a817c236134dc9c7317b32c69a5/rs/http_endpoints/public/src/lib.rs#L344)
one is created when starting the ingress watcher and it is not stored
anywhere to be cancelled later), but the cancellation token is cancelled
upon
[drop](https://github.com/dfinity/ic/blob/26d5f9d0bdca0a817c236134dc9c7317b32c69a5/rs/state_machine_tests/src/lib.rs#L1486)
in StateMachine tests resulting in errors being logged. Hence, this PR
updates the log level of ingress watcher cancellation from error to
info.