feat: [MR-580] Raise a critical error if ReplicatedState gets altered during checkpointing (#1079)
When creating a checkpoint, we serialize ReplicatedState to create a
checkpoint. We then load said checkpoint, but for execution use the
original ReplicatedState patched to get new files in pagemaps.
If these diverge, we can get a divergence between nodes continuing
execution and starting replica process.
This change creates an alert if a divergence is detected.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>