test: Add a conformance test which ensures that components only reference file paths that are present in the built image (#1053)
NODE-1458

The check is implemented as a Bazel test. It can be run manually (eg.
`bazel test //ic-os/hostos/envs/dev:dev_component_conformance_test`) and
will be included in the CI run.

A few weaknesses of the check that may be addressed later:
  - Only paths in /opt/ic are verified.
- Only constant paths are supported (without variable/command
substitution).
- Conditional references (eg. `if [ -e ... ]` ) aren't supported, an
allowlist was added to bypass these.

---------

Signed-off-by: David Frank <david.frank@dfinity.org>