feat(testnet): Use .wasm.gz for NNS canisters instead of uncompressed (#1115)
Since https://github.com/dfinity/ic/pull/952 the NNS canisters
significantly increased in size, and we now need to upgrade them using
the compressed wasm.
In https://github.com/dfinity/ic/pull/1024 the `ic-nns-init` got support
for `*.wasm.gz` files, and this PR is a follow-up, which uses the new
`ic-nns-init` functionality and keeps the canister wasm files
uncompressed.

I tested that the latest master of the IC repo can be deployed.
```
./testnet/tools/icos_deploy.sh --git-revision $(gitlab-ci/src/artifacts/newest_sha_with_disk_image.sh master) benchmarkxsmall01
```