feat(ic-admin): Let ic-admin accept an  param and let NNS release script pass it (#1154)
# Why

For the `wasm_module`, ic-admin accepts the path/url as well as its
hash, so it's less error prone for users. Since there is symmetry
between `wasm_module` and `arg`, we support the functionality for `arg`
as well.

# What

* Add `--arg_sha256` ic-admin argument
* Check `--arg_sha256` against `--arg` when both are provided, and panic
when only one of them is
* Provide the new argument to `ic-admin` within the NNS release script

# Tests

Tested locally for upgrade proposals with and without upgrade args