refactor(crypto): CRP-2559 remove CspSigVerifier (#653)
Refactors the code so that the crypto component bypasses the obsolete
`CspSigVerifier` and at the same time deletes it. This is a step towards
removing the crypto component's obsolete CSP (crypto service provider)
layer.

Closes CRP-2559 and CRP-2139.

Now `ic-crypto-ed25519` is called directly for performing batch
verification. Neither `ic-crypto-internal-csp` nor
`ic-crypto-internal-basic-sig-ed25519` are involved any longer and all
unused code is removed.

So far, in `CspSigVerifier::verify_batch`, `self.verify` was called
iteratively for algorithms different from `Ed25519`. Because batch
verification is not needed and in production also never called for
anything else than `Ed25519`, the respective logic was removed, and a
respective error ("Only Ed25519 is supported in batched basic sig
verification.") is returned if batch verification is called if the
algorithm is not `Ed25519`.

Test coverage on the crypto component level was already very good. For
the parts that were tested in the intermediate levels (which tests are
removed as part of this PR), new crypto component integration tests were
created.