feat(ckbtc): add an integration test to ic-btc-kyt (#1150)
XC-151: Add a pocket-ic based integration test to the ic-btc-kyt
package.

However, bazel compilation of ic-btc-kyt would give unknown imports
(from secp256k1 C library) in wasm file. To fix this problem, the
secp256k1-sys crate must be compiled with
`--cfg=rust_secp_no_symbol_renaming` flag. Further more, the flag needs
to be pinned to a specific version (or better yet, a specific target
platform, but this seems not possible at the moment) to avoid adversely
affecting other use of this package. So we must be cautious about
updating the version of secp256k1-sys in the future.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>